requirements.txt
flask-ngrok
flask
tensorflow
dlib
opencv-python
os
numpy
Pillow
matplotlib